+++ VisualChatGPTStudioShared/ToolWindows/TerminalWindow/TerminalWindow.cs
@@ -44,10 +44,9 @@ namespace JeffPires.VisualChatGPTStudio.ToolWindows
         /// </summary>
         /// <param name="command">The command to send to the ChatGPT window.</param>
         /// <param name="selectedText">The selected text to be sent.</param>
-        /// <param name="removeCodeTagsFromOpenAIResponses">Indicates if the code tags from OpenAI responses need to be removed from the result.</param>
-        public async System.Threading.Tasks.Task RequestToWindowAsync(string command, string selectedText, bool removeCodeTagsFromOpenAIResponses)
+        public async System.Threading.Tasks.Task RequestToWindowAsync(string command, string selectedText)
         {
-            await ((TerminalWindowControl)this.Content).RequestToWindowAsync(command, selectedText, removeCodeTagsFromOpenAIResponses);
+            await ((TerminalWindowControl)this.Content).RequestToWindowAsync(command, selectedText);
         }
     }
 }